# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShopIA API",
    "version": "1.0.0",
    "description": "Documentation API ShopIA Dashboard"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
...
    "/api/boutiques/prestataires/liste": {
      "get": {
        "summary": "Liste des prestataires de service",
        "description": "Tableau de suivi détaillé par prestataire avec toutes les métriques",
        "tags": [
          "Prestataires"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "medecin",
                "hotel",
                "restaurant",
                "salon",
                "autre"
              ],
              "default": "all"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Recherche par nom de prestataire"
          },
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month",
                "year"
              ],
              "default": "month"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prestataires récupérés avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "prestataire": {
                                "type": "string"
                              },
                              "type": {
                                "type": "string"
                              },
                              "rdv_actifs": {
                                "type": "integer"
                              },
                              "reservations": {
                                "type": "integer"
                              },
                              "taux_occupation": {
                                "type": "number"
                              },
                              "annules_non_honores": {
                                "type": "integer"
                              },
                              "bot_whatsapp": {
                                "type": "string",
                                "enum": [
                                  "actif",
                                  "inactif"
                                ]
                              },
                              "campagnes": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/boutiques/prestataires/campagnes": {
      "get": {
        "summary": "Campagnes WhatsApp des prestataires",
        "description": "Liste les campagnes WhatsApp actives et terminées des prestataires",
        "tags": [
          "Prestataires"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month",
                "year"
              ],
              "default": "month"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "active",
                "en_pause",
                "terminee"
              ],
              "default": "all"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campagnes récupérées avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "prestataire": {
                                "type": "string"
                              },
                              "campagne": {
                                "type": "string"
                              },
                              "messages_envoyes": {
                                "type": "integer"
                              },
                              "taux_ouverture": {
                                "type": "number"
                              },
                              "statut": {
                                "type": "string",
                                "enum": [
                                  "active",
                                  "en_pause",
                                  "terminee"
                                ]
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}